Modularize mutex
